requirements.txt
alpaca-trade-api==3.2.0
alpaca-py==0.26.0
numpy==1.26.4
pandas==2.2.0
pandas-ta==0.3.14b0
fastapi==0.109.2
uvicorn[standard]==0.27.0
httpx==0.26.0
uvloop==0.19.0
lightgbm==4.2.0
onnxruntime==1.17.3
onnxmltools==1.13.0
skl2onnx==1.16.0
vectorbt==0.25.2
pytest==7.4.4
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.1.0
pytz==2024.1
